fix til bitfield size calculation
