+++ src/til/section.rs
@@ -1,20 +1,15 @@
 use crate::id0::{Compiler, Id0TilOrd};
 use crate::ida_reader::{IdaGenericBufUnpack, IdaGenericUnpack};
-use crate::til::{flag, Basic, TILMacro, TILTypeInfo, TypeVariant};
+use crate::til::{flag, TILMacro, TILTypeInfo};
 use crate::IDBSectionCompression;
 use anyhow::{anyhow, ensure, Result};
 use serde::{Deserialize, Serialize};
 
-use std::collections::{HashMap, HashSet};
 use std::fmt::Debug;
 use std::io::{BufReader, Read, Write};
 use std::num::NonZeroU8;
 
 use super::function::{CCModel, CCPtrSize, CallingConvention};
-use super::r#enum::Enum;
-use super::r#struct::Struct;
-use super::union::Union;
-use super::{Type, Typedef};
 
 // TODO migrate this to flags
 pub const TIL_SECTION_MAGIC: &[u8; 6] = b"IDATIL";
@@ -682,140 +677,5 @@ impl TILSection {
             "TypeBucket compressed data is smaller then expected"
         );
         Ok(())
-    }
-}
-
-pub struct TILTypeSizeSolver<'a> {
-    section: &'a TILSection,
-    solved: HashMap<usize, u64>,
-    // HACK used to avoid infinte lopping during recursive solving
-    solving: HashSet<usize>,
-}
-
-impl<'a> TILTypeSizeSolver<'a> {
-    pub fn new(section: &'a TILSection) -> Self {
-        Self {
-            section,
-            solved: Default::default(),
-            solving: Default::default(),
-        }
-    }
-
-    // TODO make a type for type_idx and symbol_idx, accept both here
-    /// NOTE that type_idx need to be specified if not a symbol
-    pub fn type_size_bytes(&mut self, type_idx: Option<usize>, ty: &Type) -> Option<u64> {
-        assert!(self.solving.is_empty());
-        if let Some(idx) = type_idx {
-            // if cached return it
-            if let Some(solved) = self.cached(idx) {
-                return Some(solved);
-            }
-            self.solving.insert(idx);
-        }
-        let result = self.inner_type_size_bytes(ty);
-        if let Some(idx) = type_idx {
-            assert!(self.solving.remove(&idx));
-        }
-        assert!(self.solving.is_empty());
-        if let (Some(idx), Some(result)) = (type_idx, result) {
-            assert!(self.solved.insert(idx, result).is_none());
-        }
-        result
-    }
-
-    fn cached(&self, idx: usize) -> Option<u64> {
-        self.solved.get(&idx).copied()
-    }
-
-    fn inner_type_size_bytes(&mut self, ty: &Type) -> Option<u64> {
-        Some(match &ty.type_variant {
-            TypeVariant::Basic(Basic::Char) => 1,
-            // TODO what is the SegReg size?
-            TypeVariant::Basic(Basic::SegReg) => 1,
-            TypeVariant::Basic(Basic::Void) => 0,
-            TypeVariant::Basic(Basic::Unknown { bytes }) => (*bytes).into(),
-            TypeVariant::Basic(Basic::Bool) => self.section.size_bool.get().into(),
-            TypeVariant::Basic(Basic::Short { .. }) => self.section.sizeof_short().get().into(),
-            TypeVariant::Basic(Basic::Int { .. }) => self.section.size_int.get().into(),
-            TypeVariant::Basic(Basic::Long { .. }) => self.section.sizeof_long().get().into(),
-            TypeVariant::Basic(Basic::LongLong { .. }) => {
-                self.section.sizeof_long_long().get().into()
-            }
-            TypeVariant::Basic(Basic::IntSized { bytes, .. }) => bytes.get().into(),
-            TypeVariant::Basic(Basic::BoolSized { bytes }) => bytes.get().into(),
-            // TODO what's the long double default size if it's not defined?
-            TypeVariant::Basic(Basic::LongDouble) => self
-                .section
-                .size_long_double
-                .map(|x| x.get())
-                .unwrap_or(8)
-                .into(),
-            TypeVariant::Basic(Basic::Float { bytes }) => bytes.get().into(),
-            // TODO is pointer always near? Do pointer size default to 4?
-            TypeVariant::Pointer(_) => self.section.addr_size().get().into(),
-            TypeVariant::Function(_) => 0, // function type dont have a size, only a pointer to it
-            TypeVariant::Array(array) => {
-                let element_len = self.inner_type_size_bytes(&array.elem_type)?;
-                element_len * array.nelem as u64
-            }
-            TypeVariant::StructRef(ref_type)
-            | TypeVariant::UnionRef(ref_type)
-            | TypeVariant::EnumRef(ref_type)
-            | TypeVariant::Typedef(ref_type) => self.solve_typedef(ref_type)?,
-            TypeVariant::Struct(Struct { members, .. }) => {
-                let mut sum = 0u64;
-                // TODO default alignment, seems like default alignemnt is the field size
-                let align: u64 = 1;
-                for member in members {
-                    let field_size = self.inner_type_size_bytes(&member.member_type)?;
-                    let align_diff = sum % align;
-                    if align_diff != 0 {
-                        sum += align - align_diff;
-                    }
-                    sum += field_size;
-                }
-                sum
-            }
-            TypeVariant::Union(Union { members, .. }) => {
-                let mut max = 0;
-                for (_, member) in members {
-                    let size = self.inner_type_size_bytes(member)?;
-                    max = max.max(size);
-                }
-                max
-            }
-            TypeVariant::Enum(Enum { storage_size, .. }) => storage_size
-                .or(self.section.size_enum)
-                .map(|x| x.get())
-                .unwrap_or(4)
-                .into(),
-            TypeVariant::Bitfield(bitfield) => bitfield.width.into(),
-        })
-    }
-
-    fn solve_typedef(&mut self, typedef: &Typedef) -> Option<u64> {
-        let idx = match typedef {
-            Typedef::Name(name) => {
-                // NOTE missing names may indicate a external type, just return no size
-                self.section.get_name_idx(name.as_ref()?)?
-            }
-            Typedef::Ordinal(ord) => self
-                .section
-                .get_ord_idx(crate::id0::Id0TilOrd { ord: (*ord).into() })?,
-        };
-        // if cached return it
-        if let Some(solved) = self.cached(idx) {
-            return Some(solved);
-        }
-        if !self.solving.insert(idx) {
-            return None;
-        }
-        let inner_type = self.section.get_type_by_idx(idx);
-        let result = self.inner_type_size_bytes(&inner_type.tinfo);
-        self.solving.remove(&idx);
-        if let Some(result) = result {
-            assert!(self.solved.insert(idx, result).is_none());
-        }
-        result
     }
 }
